backend/pyproject.toml
[project]
name = "tradegod"
version = "0.1.0"
description = "Trading application backend"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.31.0",
    "fastapi>=0.135.3",
    "pydantic-settings>=2.13.1",
    "pydantic[email]>=2.13.0",
    "sqlalchemy>=2.0.49",
    "structlog>=25.5.0",
    "uvicorn[standard]>=0.44.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
tradegod = "tradegod.main:main"

[tool.pyright]
venvPath = "."
venv = ".venv"
pythonVersion = "3.13"
include = ["src"]
extraPaths = ["src"]

[tool.ruff]
line-length = 140

[tool.ruff.lint.per-file-ignores]
"src/tradegod/main.py" = ["E402"]

[dependency-groups]
dev = [
    "alembic>=1.18.4",
    "pytest>=9.0.3",
    "ruff>=0.15.10",
]
